4/23/2025

+++ src/Q4/Prog1999w/Cow.java
@@ -4,7 +4,7 @@ public class Cow extends Animal implements Comparable<Cow> {
     private int milk;
     private double milkPrice;
 
-    public Cow(int id, double weight, int cornCobs, int hayBales, int beans, int oats, int milk, int milkPrice) {
+    public Cow(int id, double weight, int cornCobs, int hayBales, int beans, int oats, int milk, double milkPrice) {
         super(weight, cornCobs, hayBales, beans, oats);
         this.milk = milk;
         this.milkPrice = milkPrice;
